# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.createReviewRequest:
    properties:
      comment:
        type: string
      listingId:
        type: string
      rating:
        type: integer
    type: object
  handler.createShipmentRequest:
    properties:
      courierName:
        type: string
      method:
        type: string
      trackingNumber:
        type: string
    type: object
...
  handler.updateShipmentStatusRequest:
    properties:
      note:
        type: string
      status:
        type: string
    type: object
...
  models.PublicUser:
    properties:
      avatarUrl:
        type: string
      dormBuilding:
        type: string
      id:
        type: string
      name:
        type: string
      trustScore:
        type: integer
    type: object
  models.Review:
    properties:
      comment:
        type: string
      createdAt:
        type: string
      id:
        type: string
      listingId:
        type: string
      rating:
        description: 1-5
        type: integer
      revieweeId:
        type: string
      reviewer:
        $ref: '#/definitions/models.PublicUser'
      reviewerId:
        type: string
    type: object
  models.Shipment:
    properties:
      conversationId:
        type: string
      courierName:
        type: string
      createdAt:
        type: string
      events:
        items:
          $ref: '#/definitions/models.ShipmentEvent'
        type: array
      id:
        type: string
      method:
        $ref: '#/definitions/models.ShipmentMethod'
      status:
        $ref: '#/definitions/models.ShipmentStatus'
      trackingNumber:
        type: string
      updatedAt:
        type: string
    type: object
  models.ShipmentEvent:
    properties:
      createdAt:
        type: string
      id:
        type: string
      note:
        type: string
      shipmentId:
        type: string
      status:
        $ref: '#/definitions/models.ShipmentStatus'
    type: object
  models.ShipmentMethod:
    enum:
    - pickup
    - delivery
    type: string
    x-enum-varnames:
    - ShipmentMethodPickup
    - ShipmentMethodDelivery
  models.ShipmentStatus:
    enum:
    - pending
    - shipped
    - completed
    - cancelled
    type: string
    x-enum-varnames:
    - ShipmentPending
    - ShipmentShipped
    - ShipmentCompleted
    - ShipmentCancelled
...
host: localhost:8080
info:
  contact:
    email: tnswnb@gmail.com
    name: DormMarket Support
  description: REST API สำหรับแพลตฟอร์มตลาดซื้อขายของมือสองในหอ/มหาลัย (DormMarket)
  license:
    name: MIT
  title: DormMarket API
  version: "1.0"
paths:
...
  /api/conversations/{id}/shipment:
    get:
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Shipment'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: ดูข้อมูลการจัดส่ง
      tags:
      - shipments
    post:
      consumes:
      - application/json
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: ข้อมูลการจัดส่ง
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createShipmentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: สร้างข้อมูลการจัดส่ง (ผู้ขายเท่านั้น)
      tags:
      - shipments
  /api/conversations/{id}/shipment/status:
    patch:
      consumes:
      - application/json
      description: 'สถานะที่รองรับ: pending, shipped, completed, cancelled'
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: สถานะใหม่ + หมายเหตุ
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.updateShipmentStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Shipment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: อัปเดตสถานะการจัดส่ง (ผู้ขายเท่านั้น)
      tags:
      - shipments
...
  /api/listings/{id}/can-review:
    get:
      description: ใช้โชว์/ซ่อนปุ่ม "เขียนรีวิว" ฝั่ง frontend
      parameters:
      - description: Listing ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: เช็คว่ารีวิวประกาศนี้ได้หรือยัง
      tags:
      - reviews
...
  /api/reviews:
    post:
      consumes:
      - application/json
      description: รีวิวได้เฉพาะประกาศที่สถานะ sold แล้วและเป็นผู้ซื้อในสนทนานั้นจริง
      parameters:
      - description: ข้อมูลรีวิว
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Review'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: เขียนรีวิวผู้ขาย
      tags:
      - reviews
  /api/users/{id}/reviews:
    get:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Review'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: รีวิวทั้งหมดของผู้ใช้คนหนึ่ง (public)
      tags:
      - reviews
securityDefinitions:
  BearerAuth:
    description: พิมพ์ "Bearer" ตามด้วยเว้นวรรคแล้วตามด้วย JWT token เช่น "Bearer
      eyJhbGc..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"